backend-orchestrator/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
google-generativeai==0.8.3
httpx==0.27.2
asyncpg==0.31.0
pgvector==0.3.6
redis==5.0.8
python-multipart==0.0.22
